cmd: Add snapshot subcommand.
This subcommand groups all gadgets which output snapshot information like
process-collector or socket-collector.
