+++ cmd/kubectl-gadget/main.go
@@ -19,6 +19,7 @@ import (
 
 	"github.com/spf13/cobra"
 
+	"github.com/kinvolk/inspektor-gadget/cmd/kubectl-gadget/snapshot"
 	"github.com/kinvolk/inspektor-gadget/cmd/kubectl-gadget/top"
 	"github.com/kinvolk/inspektor-gadget/cmd/kubectl-gadget/trace"
 
@@ -36,6 +37,7 @@ var rootCmd = &cobra.Command{
 func init() {
 	utils.FlagInit(rootCmd)
 
+	rootCmd.AddCommand(snapshot.SnapshotCmd)
 	rootCmd.AddCommand(top.TopCmd)
 	rootCmd.AddCommand(trace.TraceCmd)
 }
