+++ integration/integration_test.go
@@ -468,7 +468,7 @@ func TestProcessCollector(t *testing.T) {
 		waitUntilTestPodReadyCommand(ns),
 		{
 			name:           "Run process-collector gadget",
-			cmd:            fmt.Sprintf("$KUBECTL_GADGET process-collector -n %s", ns),
+			cmd:            fmt.Sprintf("$KUBECTL_GADGET snapshot process -n %s", ns),
 			expectedRegexp: fmt.Sprintf(`%s\s+test-pod\s+test-pod\s+nginx\s+\d+`, ns),
 		},
 		deleteTestNamespaceCommand(ns),
@@ -663,7 +663,7 @@ func TestSocketCollector(t *testing.T) {
 		waitUntilTestPodReadyCommand(ns),
 		{
 			name:           "Run socket-collector gadget",
-			cmd:            fmt.Sprintf("$KUBECTL_GADGET socket-collector -n %s", ns),
+			cmd:            fmt.Sprintf("$KUBECTL_GADGET snapshot socket -n %s", ns),
 			expectedRegexp: fmt.Sprintf(`%s\s+test-pod\s+TCP\s+0\.0\.0\.0`, ns),
 		},
 		deleteTestNamespaceCommand(ns),
